+++ Sources/MeerkatKit/ShakeResponderBridge.swift
@@ -1,0 +1,15 @@
+import SwiftUI
+
+struct ShakeResponderBridge: View {
+    let onShake: () -> Void
+
+    var body: some View {
+        #if os(iOS)
+        ShakeResponderView(onShake: onShake)
+            .frame(width: 0, height: 0)
+            .accessibilityHidden(true)
+        #else
+        EmptyView()
+        #endif
+    }
+}